requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==3.0.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
mysqlclient==2.2.5
psycopg2-binary==2.9.10
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
whitenoise==6.9.0
python-dateutil==2.9.0

